gimp#14544: building GIMP 3.0.4 fails with old librsvg.
Thanks to Luca Bonissi for the initial patch, from which this commit is
nearly a copy. Though I do check for librsvg 2.46 and not 2.45.90 (doc
says rsvg_handle_render_document() is available since version 2.46).

Note that I have not tested this patch with an older librsvg myself. It
looks sane and I am assuming that the contributor did verify it works as
expected on older versions of the library.