images: update the image splash requirements.
For splash images made to be used for stable releases, we now add a
requirement of adding a text layer containing only the text "RC".

First we will now avoid counting the RCs in the splash image because if
we forget, we would end up with the same issue as during 3.0.0 RC
releases, where one RC splash image had the wrong number!
We could obviously edit this number through scripting but it would force
us to install the correct font on CI at every release, and the font may
change all the time so we'd always have to tweak the build scripts,
which is counter-productive (the goal is to automatize things, not
forcing us on manual tweaking at every splash change). Considering that
a RC is not as a big deal as a stable version, having a generic "RC"
seems fine.

Second, this will allow us not to have to edit the splash image
unnecessarily, within the RC period (if we change the image), but even
at actual stable release time!