images: update splash image.
This is the second image contributed by Mark McCaughrean.
Cf. #11.

It was also published by Mark as CC by-sa 2.0:
https://www.flickr.com/photos/markmccaughrean/53236144002/in/dateposted/
Then further edited for being used as a splash image.

This is a "JWST image of the Trapezium Cluster & inner Orion Nebula:
short-wavelength channel".

The rest of information about fonts used and workflow are the same as
the long-wavelength version. Cf. commit 7f74eb8.